Changed view to bind to model
