+++ Core/Models/Settings.cs
@@ -1,13 +1,8 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace CRM.Core.Models
+﻿namespace CRM.Core.Models
 {
 	public class Settings
 	{
 		public int Id { get; set; }
 		public bool FirstRun { get; set; }
-
-		[Required]
-		public EmailSettings EmailSettings { get; set; } = null!;
 	}
 }
