+++ Server/Pages/Setup.cshtml
@@ -8,45 +8,45 @@
 
 <div class="content">
 	<section class="form-section">
-		<form id="email-setup" method="post">
+		<form asp-controller="EmailSettings" method="post">
 			<h1>Setup your email server</h4>
 
 			<div class="form-line">
-				<fluent-text-field id="Name" name="Name" placeholder="@Html.DisplayNameFor(model => model.Name)" value="@Model.Name" data-val="true" appearance="filled" />
-				<span asp-validation-for="Name" class="text-danger"></span>
+				<fluent-text-field id="Settings_Name" name="Settings.Name" placeholder="@Html.DisplayNameFor(model => model.Settings.Name)" value="@Model.Settings?.Name" data-val="true" appearance="filled" />
+				<span asp-validation-for="Settings.Name" class="text-danger"></span>
 			</div>
 
 			<div class="form-line">
-				<fluent-text-field id="Address" name="Address" placeholder="@Html.DisplayNameFor(model => model.Address)" type="email" value="@Model.Address" data-val="true" appearance="filled" />
-				<span asp-validation-for="Address" class="text-danger"></span>
+				<fluent-text-field id="Settings_Address" name="Settings.Address" placeholder="@Html.DisplayNameFor(model => model.Settings.Address)" type="email" value="@Model.Settings?.Address" data-val="true" appearance="filled" />
+				<span asp-validation-for="Settings.Address" class="text-danger"></span>
 			</div>
 
 			<div class="form-line">
 				<div class="form-line-cell">
-					<fluent-text-field id="Server" name="Server" style="width: 188px;" placeholder="@Html.DisplayNameFor(model => model.Server)" value="@Model.Server" data-val="true" appearance="filled" />
-					<span asp-validation-for="Server" class="text-danger" />
+					<fluent-text-field id="Settings_Server" name="Settings.Server" style="width: 188px;" placeholder="@Html.DisplayNameFor(model => model.Settings.Server)" value="@Model.Settings?.Server" data-val="true" appearance="filled" />
+					<span asp-validation-for="Settings.Server" class="text-danger" />
 				</div>
 				
 				<div class="form-line-cell">
-					<fluent-text-field id="Port" name="Port" style="width: 94px;" placeholder="@Html.DisplayNameFor(model => model.Port)" type="number" value="@Model.Port" data-val="true" appearance="filled" />
-					<span asp-validation-for="Port" class="text-danger"></span>
+					<fluent-text-field id="Settings_Port" name="Settings.Port" style="width: 94px;" placeholder="@Html.DisplayNameFor(model => model.Settings.Port)" type="number" value="@Model.Settings?.Port" data-val="true" appearance="filled" />
+					<span asp-validation-for="Settings.Port" class="text-danger"></span>
 				</div>
 			</div>
 
 			<div class="form-line">
-				<label asp-for="SecureSocket"></label>
-				<select asp-for="SecureSocket" asp-items="Html.GetEnumSelectList<SecureSocket>()"></select>
-				<span asp-validation-for="SecureSocket" class="text-danger"></span>
+				<label asp-for="Settings.SecureSocket"></label>
+				<select asp-for="Settings.SecureSocket" asp-items="Html.GetEnumSelectList<SecureSocket>()"></select>
+				<span asp-validation-for="Settings.SecureSocket" class="text-danger"></span>
 			</div>
 
 			<div class="form-line">
-				<fluent-text-field id="Login" name="Login" placeholder="@Html.DisplayNameFor(model => model.Login)" type="email" value="@Model.Login" data-val="true" appearance="filled" />
-				<span asp-validation-for="Login" class="text-danger"></span>
+				<fluent-text-field id="Settings_Login" name="Settings.Login" placeholder="@Html.DisplayNameFor(model => model.Settings.Login)" type="email" value="@Model.Settings?.Login" data-val="true" appearance="filled" />
+				<span asp-validation-for="Settings.Login" class="text-danger"></span>
 			</div>
 
 			<div class="form-line">
-				<fluent-text-field id="Password" name="Password" placeholder="@Html.DisplayNameFor(model => model.Password)" type="password" data-val="true" appearance="filled" />
-				<span asp-validation-for="Password" class="text-danger"></span>
+				<fluent-text-field id="Settings_Password" name="Settings.Password" placeholder="@Html.DisplayNameFor(model => model.Settings.Password)" type="password" data-val="true" appearance="filled" />
+				<span asp-validation-for="Settings.Password" class="text-danger"></span>
 			</div>
 
 			<div class="form-line">
@@ -54,5 +54,4 @@
 			</div>
 		</form>
 	</section>
-	<div />
 </div>
