+++ Server/appsettings.Development.json
@@ -1,5 +1,6 @@
 {
 	"DetailedErrors": true,
+	"BaseUrl": "https://localhost:5001/",
 	"Logging": {
 		"LogLevel": {
 			"Default": "Information",
